chore: reduce duplication and have a shared screening page
